requirements/base.txt
Django==5.2.2
python-decouple==3.8
psycopg2-binary==2.9.10
djangorestframework==3.16.0
django-filter==25.1
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
django-filter==25.1
django-tenants==3.8.0
pillow==11.2.1

